fix point_in_polygon
